% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\alias{factorize}
\alias{linear_regression}
\alias{estimate}
\alias{square}
\title{Hello, world!}
\usage{
linear_regression(formula=y~x1+x2)
estimate(x,y)
factorize(x)
square(x)
}
\arguments{
\item{x, y, formula}{input value}
}
\value{
the matrix_x
}
\description{
This is an example function named 'hello'
which prints 'Hello, world!'.
}
\details{
Some useful keyboard shortcuts for package authoring:

Install Package:           'Ctrl + Shift + B'
  Check Package:             'Ctrl + Shift + E'
  Test Package:              'Ctrl + Shift + T'

value all components of linear_regression results are listed below:
x matrix
formula y ~ x1+x2
}
\examples{
x1 <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
y <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
x2 <- c(2, 10, 20, 2, 10, 20, 2, 10, 20, 2)
x <- c(x1, x2)
formula=y~x1+x2
lm.D9 <- linear_regression(formula)
lm.D90 <- estimate(x1,y)
lm.D99 = factorize(x)
square(x)
}
