% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{linear_regression}
\alias{linear_regression}
\alias{factorize}
\alias{estimate}
\title{Hello, world!}
\usage{
linear_regression(formula)
estimate(x,y)
factorize(x)
}
\arguments{
\item{x, y, formula}{input value}
}
\value{
the matrix_x
}
\description{
This is an example function named 'hello'
which prints 'Hello, world!'.
}
\details{
Some useful keyboard shortcuts for package authoring:

Install Package:           'Ctrl + Shift + B'
  Check Package:             'Ctrl + Shift + E'
  Test Package:              'Ctrl + Shift + T'

value all components of linear_regression results are listed below:
x matrix
formula y ~ x1+x2
}
\examples{
x1=c(1,2,3,4,5,6,7,8)
y=c(23,24,26,37,38,25,36,40)
x2=c(23,32,34,20,24,56,34,24)
x=c("M","F","F","U","M","F","F","U")
formula=y~x1+x2
lm.D9 <- linear_regression(formula)
lm.D90 <- estimate(x1,y)
lm.D99 = factorize(x)

}
