\name{LR_addis-package}
\alias{LR_addis-package}
\alias{LR_addis}
\docType{package}
\title{
\packageTitle{LR_addis}
}
\description{
\packageDescription{LR_addis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LR_addis}
\packageIndices{LR_addis}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LR_addis}

Maintainer: \packageMaintainer{LR_addis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
